'修改学分'

+++ src/main/java/com/itcast/tpms/exp/CourseExpList.java
@@ -12,12 +12,12 @@ public class CourseExpList {
 
     private String moduleName;
     private List<CourseExp> courseExps = new ArrayList<>();
-    private Integer totalCredit;
+    private Float totalCredit;
     private Integer totalClassHour;
     private Integer totalCourse;
 
     public void init() {
-        totalCredit = 0;
+        totalCredit = 0.f;
         totalClassHour = 0;
         totalCourse = 1;
         for (CourseExp courseExp : courseExps) {
