+++ src/main/java/com/itcast/tpms/exp/OneLevelList.java
@@ -12,12 +12,12 @@ import java.util.List;
 public class OneLevelList {
 
     private List<TwoLevelList> oneLevels = new ArrayList<>();
-    private Integer totalCredit;
+    private Float totalCredit;
     private Integer totalClassHour;
     private Integer totalCourse;
 
     public void init() {
-        totalCredit = 0;
+        totalCredit = 0.f;
         totalClassHour = 0;
         totalCourse = 1;
         for (TwoLevelList twoLevelList : oneLevels) {
